pyproject.toml
[project]
name = "otolith-age-prediction"
version = "0.1.0"
description = "Few-shot learning for cod otolith age determination using CLIP"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Joao DSM"}
]

dependencies = [
    # Core ML - PyTorch with MPS support for Apple Silicon
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # HuggingFace Transformers for CLIP and SigLIP2
    "transformers>=4.49.0",  # Required for SigLIP2 support

    # ML utilities
    "scikit-learn>=1.3.0",
    "numpy>=2.3.3",
    "pandas>=2.0.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Image processing
    "pillow>=10.3.0,<11.0.0",

    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "ipywidgets>=8.0.0",
]

#[project.scripts]
#extract-features = "scripts.extract_features:main"
#train-shallow = "scripts.train_shallow:main"
#run-experiment = "scripts.run_experiment:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
